Fixes after review: - rename validateRequestParameters to createRequestParameters - separate query building functionality from Dao into utils class
