+++ src/test/java/com/anabol/movieland/dao/jdbc/utils/QueryBuilderTest.java
@@ -1,15 +1,17 @@
-package com.anabol.movieland.dao.jdbc;
+package com.anabol.movieland.dao.jdbc.utils;
 
 import com.anabol.movieland.web.utils.RequestParameters;
 import com.anabol.movieland.web.utils.SortDirection;
 import org.junit.Test;
+
+import static com.anabol.movieland.dao.jdbc.utils.QueryBuilder.addOrder;
 import static org.junit.Assert.assertEquals;
 
-public class JdbcMovieDaoTest {
+public class QueryBuilderTest {
     @Test
     public void testAddOrder() {
         RequestParameters requestParameters = new RequestParameters("column1", SortDirection.DESC);
-        String query = JdbcMovieDao.addOrder("SELECT * FROM table", requestParameters);
+        String query = addOrder("SELECT * FROM table", requestParameters);
         assertEquals("SELECT * FROM table ORDER BY column1 DESC", query);
     }
